requirements.txt.txt
streamlit
opencv-contrib-python-headless
rembg
Pillow
numpy
requests